feat(vsc): reorder template UI layout
